Add 3 new EventHandlers and the first transactions

+++ AzzyBot-Next/Services/DiscordBotServiceHost.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Transactions;
 using AzzyBot.Commands;
 using AzzyBot.Commands.Converters;
 using AzzyBot.Database;
@@ -34,12 +32,14 @@ internal sealed class DiscordBotServiceHost : IHostedService
     private readonly ILoggerFactory _loggerFactory;
     private readonly IServiceProvider _serviceProvider;
     private readonly AzzyBotSettingsRecord _settings;
+    private readonly IDbContextFactory<DatabaseContext> _dbContextFactory;
     internal readonly DiscordShardedClient _shardedClient;
     private DiscordBotService? _botService;
 
-    public DiscordBotServiceHost(AzzyBotSettingsRecord settings, ILogger<DiscordBotServiceHost> logger, ILoggerFactory loggerFactory, IServiceProvider serviceProvider)
+    public DiscordBotServiceHost(AzzyBotSettingsRecord settings, IDbContextFactory<DatabaseContext> dbContextFactory, ILogger<DiscordBotServiceHost> logger, ILoggerFactory loggerFactory, IServiceProvider serviceProvider)
     {
         _settings = settings;
+        _dbContextFactory = dbContextFactory;
         _logger = logger;
         _loggerFactory = loggerFactory;
         _serviceProvider = serviceProvider;
@@ -147,12 +147,15 @@ internal sealed class DiscordBotServiceHost : IHostedService
         _shardedClient.ClientErrored += ShardedClientErroredAsync;
         _shardedClient.GuildCreated += ShardedClientGuildCreatedAsync;
         _shardedClient.GuildDeleted += ShardedClientGuildDeletedAsync;
+        _shardedClient.GuildDownloadCompleted += ShardedClientGuildDownloadCompletedAsync;
     }
+
     private void UnregisterEventHandlers()
     {
         _shardedClient.ClientErrored -= ShardedClientErroredAsync;
         _shardedClient.GuildCreated -= ShardedClientGuildCreatedAsync;
         _shardedClient.GuildDeleted -= ShardedClientGuildDeletedAsync;
+        _shardedClient.GuildDownloadCompleted -= ShardedClientGuildDownloadCompletedAsync;
     }
 
     private async Task CommandErroredAsync(CommandsExtension c, CommandErroredEventArgs e)
@@ -184,7 +187,7 @@ internal sealed class DiscordBotServiceHost : IHostedService
     {
         _logger.GuildCreated(e.Guild.Name);
 
-        await using DatabaseContext context = new();
+        await using DatabaseContext context = await _dbContextFactory.CreateDbContextAsync();
         await using IDbContextTransaction transaction = await context.Database.BeginTransactionAsync();
 
         try
@@ -197,31 +200,84 @@ internal sealed class DiscordBotServiceHost : IHostedService
         catch (Exception ex) when (ex is DbUpdateException || ex is DbUpdateConcurrencyException)
         {
             _logger.DatabaseTransactionFailed(ex);
+            await transaction.RollbackAsync();
         }
 
-        await e.Guild.Owner.SendMessageAsync("Thank you for adding me to your guild! Before you can make use of me, you have to set my settings first.\n\nPlease use the command `settings set` for this.\nOnly you are able to execute this command right now.");
+        await e.Guild.Owner.SendMessageAsync("Thank you for adding me to your server! Before you can make use of me, you have to set my settings first.\n\nPlease use the command `settings set` for this.\nOnly you are able to execute this command right now.");
     }
 
     private async Task ShardedClientGuildDeletedAsync(DiscordClient c, GuildDeleteEventArgs e)
     {
         _logger.GuildDeleted(e.Guild.Name);
 
-        await using DatabaseContext context = new();
+        await using DatabaseContext context = await _dbContextFactory.CreateDbContextAsync();
         await using IDbContextTransaction transaction = await context.Database.BeginTransactionAsync();
 
         try
         {
-            GuildsEntity guild = await context.Guilds.SingleAsync(g => g.UniqueId == e.Guild.Id);
-            AzuraCastEntity azura = await context.AzuraCast.SingleAsync(a => a.GuildId  == guild.Id);
-            AzuraCastChecksEntity checks = await context.AzuraCastChecks.SingleAsync(c => c.AzuraCastId == azura.Id);
+            GuildsEntity? guild = null;
+            AzuraCastEntity? azura = null;
+            AzuraCastChecksEntity? checks = null;
 
-            int guildId = guild.Id;
-            int azuraId = azura.Id;
-            int checksId = checks.Id;
+            try
+            {
+                guild = await context.Guilds.SingleAsync(g => g.UniqueId == e.Guild.Id);
+                azura = await context.AzuraCast.SingleAsync(a => a.GuildId == guild.Id);
+                checks = await context.AzuraCastChecks.SingleAsync(c => c.AzuraCastId == azura.Id);
+            }
+            catch (InvalidOperationException)
+            { }
 
-            context.AzuraCastChecks.Remove(checks);
-            context.AzuraCast.Remove(azura);
-            context.Guilds.Remove(guild);
+            if (checks is not null)
+                context.AzuraCastChecks.Remove(checks);
+
+            if (azura is not null)
+                context.AzuraCast.Remove(azura);
+
+            if (guild is not null)
+                context.Guilds.Remove(guild);
+
+            await context.SaveChangesAsync();
+
+            await transaction.CommitAsync();
+        }
+        catch (Exception ex) when (ex is DbUpdateException || ex is DbUpdateConcurrencyException)
+        {
+            _logger.DatabaseTransactionFailed(ex);
+            await transaction.RollbackAsync();
+        }
+
+        await e.Guild.Owner.SendMessageAsync("I am sorry for being not used anymore on this server. I removed every piece of data of your server from my database and hope you will find use of me again!");
+    }
+
+    private async Task ShardedClientGuildDownloadCompletedAsync(DiscordClient c, GuildDownloadCompletedEventArgs e)
+    {
+        await using DatabaseContext context = await _dbContextFactory.CreateDbContextAsync();
+        await using IDbContextTransaction transaction = await context.Database.BeginTransactionAsync();
+
+        try
+        {
+            List<GuildsEntity> guilds = await context.Guilds.ToListAsync();
+
+            foreach (DiscordGuild guild in e.Guilds.Values)
+            {
+                if (guilds.Count > 0)
+                {
+                    foreach (GuildsEntity entity in guilds)
+                    {
+                        if (guild.Id == entity.UniqueId)
+                            continue;
+
+                        await context.Guilds.AddAsync(new() { UniqueId = guild.Id });
+                    }
+                }
+                else
+                {
+                    await context.Guilds.AddAsync(new() { UniqueId = guild.Id });
+                }
+            }
+
+            await context.SaveChangesAsync();
 
             await transaction.CommitAsync();
         }
@@ -229,8 +285,6 @@ internal sealed class DiscordBotServiceHost : IHostedService
         {
             _logger.DatabaseTransactionFailed(ex);
         }
-
-        await e.Guild.Owner.SendMessageAsync("I am sorry for being not used anymore on this server. I removed every piece of data of your server from my database.");
     }
 
     private async Task ShardedClientErroredAsync(DiscordClient c, ClientErrorEventArgs e)
